feat(slider): split slider into two components
