+++ src/components/common/definitions/defineAllComponents.ts
@@ -23,6 +23,7 @@ import IgcNavbarComponent from '../../navbar/navbar';
 import IgcRadioGroupComponent from '../../radio-group/radio-group';
 import IgcRadioComponent from '../../radio/radio';
 import IgcRippleComponent from '../../ripple/ripple';
+import IgcRangeSliderComponent from '../../slider/range-slider';
 import IgcSliderComponent from '../../slider/slider';
 import { defineComponents } from './defineComponents';
 
@@ -53,6 +54,7 @@ const allComponents: CustomElementConstructor[] = [
   IgcRadioGroupComponent,
   IgcRippleComponent,
   IgcSliderComponent,
+  IgcRangeSliderComponent,
 ];
 
 export const defineAllComponents = () => {
